Out with the old.. in with the, soon-to-be-old?
I'm going to be reworking my shell and how I store my config files. This
is the starting point of that rework and is the current state of my
dotfiles.

Main difference from 1.2.0 and this current 2.0.0 is that I've swapped
to using WSL.. so all of the DE-specific configurations have been
removed.